requirements.txt
azure-core               1.32.0
azure-storage-blob       12.24.0
certifi                  2024.12.14
cffi                     1.17.1
charset-normalizer       3.4.1
contourpy                1.3.1
cryptography             44.0.0
cycler                   0.12.1
filelock                 3.16.1
fonttools                4.55.3
fsspec                   2024.12.0
idna                     3.10
isodate                  0.7.2
Jinja2                   3.1.5
kiwisolver               1.4.8
MarkupSafe               3.0.2
matplotlib               3.10.0
mpmath                   1.3.0
networkx                 3.4.2
numpy                    2.2.1
nvidia-cublas-cu12       12.4.5.8
nvidia-cuda-cupti-cu12   12.4.127
nvidia-cuda-nvrtc-cu12   12.4.127
nvidia-cuda-runtime-cu12 12.4.127
nvidia-cudnn-cu12        9.1.0.70
nvidia-cufft-cu12        11.2.1.3
nvidia-curand-cu12       10.3.5.147
nvidia-cusolver-cu12     11.6.1.9
nvidia-cusparse-cu12     12.3.1.170
nvidia-nccl-cu12         2.21.5
nvidia-nvjitlink-cu12    12.4.127
nvidia-nvtx-cu12         12.4.127
opencv-python            4.10.0.84
packaging                24.2
pandas                   2.2.3
pillow                   11.1.0
pip                      24.0
psutil                   6.1.1
py-cpuinfo               9.0.0
pycparser                2.22
pyparsing                3.2.1
python-dateutil          2.9.0.post0
pytz                     2024.2
PyYAML                   6.0.2
requests                 2.32.3
scipy                    1.15.0
seaborn                  0.13.2
setuptools               75.8.0
six                      1.17.0
sympy                    1.13.1
torch                    2.5.1
torchvision              0.20.1
tqdm                     4.67.1
triton                   3.1.0
typing_extensions        4.12.2
tzdata                   2024.2
ultralytics              8.3.58
ultralytics-thop         2.0.13
urllib3                  2.3.0

